+++ game.py
@@ -5,6 +5,15 @@ import sys,time,os
 from itertools import repeat
 from pygame import mixer
 
+def resource_path(relative_path):
+    try:
+    # PyInstaller creates a temp folder and stores path in _MEIPASS
+        base_path = sys._MEIPASS
+    except Exception:
+        base_path = os.path.abspath(".")
+
+    return os.path.join(base_path, relative_path)
+
 pygame.init()
  
 white = (255, 255, 255)
@@ -22,9 +31,13 @@ lightgrey =(224,224,224)
 
 dis_width = 900
 dis_height = 300
-background = pygame.image.load('abstract.jpg')
-blackbg = pygame.image.load('absbg.jpg')   
-icon = pygame.image.load('cube.png')
+
+asset_url1 = 'images/abstract.jpg'
+asset_url2 = 'images/absbg.jpg'
+asset_url3 = 'images/cube.png'
+background = pygame.image.load(asset_url1)
+blackbg = pygame.image.load(asset_url2)   
+icon = pygame.image.load(asset_url3)
  
 dis = pygame.display.set_mode((dis_width, dis_height))
 pygame.display.set_caption('Jump Game')
@@ -38,7 +51,6 @@ y=240
 speed = 5
 jumpcount = 8
 isjump = False
-parttime=0
 
 xobs1 = 900
 xobs2 = 1300
@@ -51,8 +63,9 @@ score = 0
 fontsize = 0
 life = 50
 
-font = pygame.font.SysFont("Arial", 18)
+
 score_font = pygame.font.Font('freesansbold.ttf', 25)
+fpsfont = pygame.font.Font('freesansbold.ttf', 15)
 
 class particle():
     def __init__(self,x1,y1):
@@ -71,8 +84,6 @@ class particle():
             self.x1 +=self.x1_vel
             self.y1 +=self.y1_vel
             pygame.draw.rect(dis,orange,[int(self.x1),int(self.y1),abs(int(self.radious)),abs(int(self.radious)+1)])
-            #pygame.draw.circle(dis,yellow,(int(self.x1-3),int(self.y1-3)),abs(int(self.radious)-2))
-            #pygame.draw.circle(dis,red,(int(self.x1),int(self.y1)),abs(int(self.radious)))
             
 particles=[]
 
@@ -104,8 +115,8 @@ class dparticle():
     def __init__(self,dx1,dy1):
         self.dx1=dx1
         self.dy1=dy1
-        self.dx1_vel = random.randrange(-1,6)*0.5
-        self.dy1_vel = random.randrange(1,6)*0.5
+        self.dx1_vel = random.randrange(-6,6)*2
+        self.dy1_vel = random.randrange(1,6)*2
         self.dlifetime = 0
         self.dradious = 10
         
@@ -116,11 +127,7 @@ class dparticle():
         if self.dlifetime<37:
             self.dx1 +=self.dx1_vel
             self.dy1 +=self.dy1_vel
-            #pygame.draw.circle(dis,black,(int(self.x1),int(self.y1)),abs(int(self.radious)-1))
-            #pygame.draw.circle(dis,yellow,(int(self.x1-3),int(self.y1-3)),abs(int(self.radious)-2))
-            #pygame.draw.circle(dis,red,(int(self.dx1-5),int(self.dy1)),abs(int(self.dradious)+7))
             pygame.draw.rect(dis,grey,[int(self.dx1),int(self.dy1),abs(int(self.dradious)),abs(int(self.dradious))])
-            pygame.draw.rect(dis,lightgrey,[int(self.dx1+3),int(self.dy1+7),abs(int(self.dradious)+2),abs(int(self.dradious)+2)])
             
 dparticles=[]
 
@@ -143,11 +150,6 @@ def highscore1():
     value = score_font.render("Highscore: " + str(highscore), True, white)
     dis.blit(value, [140, 10])
     
-def update_fps():
-	fps = str(int(clock.get_fps()))
-	fps_text = font.render(fps, 1, pygame.Color("white"))
-	return fps_text
-
 def retry():
     global score1
     global fontsize
@@ -157,6 +159,11 @@ def retry():
     dis.blit(blackbg, (0,0))
     score1()
     highscore1()
+
+def update_fps():
+	fps = str(int(clock.get_fps()))
+	fps_text = fpsfont.render(fps, 2, pygame.Color("white"))
+	return fps_text
 
 def obstacle1():
     global score
@@ -188,8 +195,6 @@ def obstacle3():
     xobs3 -= 5
     pygame.draw.polygon(dis, orange, [(xobs3,10),(xobs3+70,10),(xobs3+35,70)])
     pygame.draw.polygon(dis, white, [(xobs3,10),(xobs3+70,10),(xobs3+35,70)],1)
-    #pygame.draw.polygon(dis, blue, [(xobs3+250,10),(xobs3+320,10),(xobs3+285,50)])
-    #pygame.draw.polygon(dis, white, [(xobs3+250,10),(xobs3+320,10),(xobs3+285,50)],1)
     if xobs3<0:
         score += 1
         xobs3 = 1000
@@ -232,16 +237,17 @@ while loop:
     clr = (r,g,b)
 
     score = 0
-    pygame.mixer.music.load('bgmusic.wav')
+    pygame.mixer.music.load('bgmusic.mp3')
     pygame.mixer.music.play(0)
     soundstime = 0
     soundstime1 = 0
     
-    with open("hiscore11.txt", "r") as f:
-            highscore = f.read()
+    with open("highscore.txt", "r") as f:
+        highscore = f.read()
     while running is True:
-        while game_over == True:
-            with open("hiscore11.txt", "w") as f:
+        while game_over is True:
+            #print('game over')
+            with open("highscore.txt", "w") as f:
                 f.write(str(highscore))
             retry()
             colordur += 1
@@ -299,17 +305,12 @@ while loop:
                 isjump = False
                 jumpcount = 8
 
-        if score>=int(highscore):
-            highscore=score
+        #background---------------------------------------------------------------------------
 
         #dis.blit(background, (0,0))
         #pygame.draw.rect(dis,black,(0,0,900,300))
         dis.blit(blackbg,next(offset))
-        dis.blit(update_fps(), (800,20))
-        z=50
-        #while z<900:
-            #pygame.draw.rect(dis,white,[z,0,3,290])
-            #z +=50
+        dis.blit(update_fps(), (830,20))
         k = random.randint(20,850)
         l = random.randint(15,270)
         orparticles.append(orparticle(k,l))
@@ -317,72 +318,84 @@ while loop:
             orparticle_.draw(dis)
         score1()
         highscore1()
+
+        #player-------------------------------------------------------------------------------
         pygame.draw.rect(dis,clr,[0,0,9000,10])
         pygame.draw.rect(dis,clr,[0,290,9000,10])
         pygame.draw.rect(dis,red,[int(x),int(y),50,50])
         pygame.draw.rect(dis,red,[int(x),int(y-40),50,5])
-        pygame.draw.rect(dis,green,[int(x),int(y-40),life,5])
+        pygame.draw.rect(dis,green,[int(x),int(y-40),int(life),5])
         pygame.draw.rect(dis,white,[int(x),int(y-40),50,5],1)
         
         if y==240:
             particles.append(particle(x+5,y+30))
             for particle_ in particles:
                 particle_.draw(dis)
+
+        #empodia-------------------------------------------------------------------------------
+
         obstacle1()
         obstacle2()
         obstacle3()
         obstacle4()
-        #sygkroysh me empodia
+
+        #sygkroysh me empodia------------------------------------------------------------------
 
         distance = math.sqrt((math.pow(x - xobs1, 2)) + (math.pow(y - (280-h), 2)))
         if distance < 42:
+            print('game over1')
             soundstime += 1
             isjump = False
             speed = 0
             speedobs = 0
-            pygame.mixer.music.load('losingsound.wav')
-            pygame.mixer.music.play(0)
+            #pygame.mixer.music.load('losingsound.wav')
+            #pygame.mixer.music.play(0)
             if soundstime>=5:
+                print('game over1-2')
                 game_over = True
 
         distance1 = math.sqrt((math.pow(x - xobs2, 2)) + (math.pow(y - 240, 2)))
         if distance1 < 57:
+            print('game over2')
             soundstime += 1
             speed = 0
             isjump = False
             speedobs = 0
-            pygame.mixer.music.load('losingsound.wav')
-            pygame.mixer.music.play(0)
+            #pygame.mixer.music.load('losingsound.wav')
+            #pygame.mixer.music.play(0)
             if soundstime>=5:
+                print('game over2-2')
                 game_over = True
 
         distance2 = math.sqrt((math.pow(x - xobs3, 2)) + (math.pow(y - 90, 2)))
         if distance2 < 45:
+            print('game over3')
             lostsound = mixer.Sound('losingsound.wav')
             lostsound.play()
             life -= 0.2
             offset = shake(20)
             oldobs3 = xobs3
-            if parttime<=10:
-                parttime+=1
+            if soundstime <10:
                 dparticles.append(dparticle(oldobs3,50))
                 for dparticle_ in dparticles:
                     dparticle_.draw(dis)
-            else:
-                parttime=0
+            soundstime1 += 1
+            if soundstime1>=10:
+                xobs3 = 1000
+                soundstime1 = 0
+        
+        #-----------------------------------------------------------------------------------------
         
         if life <= 0:
+            print('game over')
             game_over = True
-        
-        #if xobs1<=0:
-            #h = random.randrange(40,80)
 
         if score >= 10:
                 speedobs += 0.0002
         if score>=int(highscore):
             highscore=score
 
-        #allagh xromaton 
+        #allagh xromaton--------------------------------------------------------------------------
         
         if 5<=score<20:#potokali
             clr = (r,g,b)
@@ -433,6 +446,4 @@ while loop:
                 b-=1
             h = 85
         pygame.display.update()
-        clock.tick(60+2*speedobs)
-        
-        
+        clock.tick(60+3*(speedobs))
