--- hiscore11.txt
@@ -1,1 +1,0 @@
-185
